wgtypes: Add Remove flag to allowed IPs
[1] adds the WGALLOWEDIP_F_REMOVE_ME flag to WireGuard's Netlink
API which, in the same way that WGPEER_F_REMOVE_ME allows a user to
remove a single peer from a WireGuard device's configuration, allows a
user to remove an ip from a peer's set of allowed ips. This capability
was subsequently ported to wireguard-go as well.

Add support for this feature to wgctrl-go, allowing clients to
incrementally remove allowed IPs on a peer like so:

wgtypes.Config{
	Peers: []wgtypes.PeerConfig{
		{
			PublicKey: peerKey,
			AllowedIPs: []wgtypes.AllowedIPConfig{
				{
					IPNet:  ip,
					Remove: true,
				},
			},
		},
	},
}

[1]: https://lore.kernel.org/netdev/20250517192955.594735-2-jordan@jrife.io/

Signed-off-by: Jordan Rife <jordan@jrife.io>