Add shim client to simulate allowed IP removal
Direct allowed IP removal is a new feature, so its availability is
limited for now. Clients who want to take advantage of this capability
would need to know ahead of time if WireGuard on their system supports
it or probe to see if they're running on a system that supports it. To
ease the transition, add the WithShim option for clients:

	c, err := wgctrl.New(wgctrl.WithShim)

WithShim wraps internal clients with a shim client that probes to see
if direct IP removal is supported on their system. If not, the shim
client emulates the effect by assigning IPs to a dummy peer then
removing that peer. At some point in the future, this option should no
longer be necessary once the feature becomes commonplace.

In my case, I plan to use WithShim to simplify Cilium's WireGuard
orchestration logic.

Signed-off-by: Jordan Rife <jordan@jrife.io>